recommender/requirements.txt
streamlit
numpy
scikit-learn
pillow